requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0.4
django-rest-auth==0.9.5
djangorestframework==3.11.0
Pillow==7.0.0
psycopg2==2.8.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
